Interrupt handler fix: Remove printf to prevent segment register corruption
- Fixed trapframe layout (DS at offset 0, pusha at offset 16)
- Removed printf from keyboard interrupt handler
- Wrapped printf in do_irq_handler with #ifdef DEBUG_IRQ_PRINT
- Added TRAPFRAME_LAYOUT_FIX.md documentation
- Added auto-generated task_offsets.s
- System now boots without TRAP 26 errors

Tested: Keyboard input works in LVGL interface
Status: Ready for merge

+++ zhwh_os/calc_task_offset.c
@@ -8,7 +8,7 @@ typedef uint32_t pid_t;
 typedef uint32_t uid_t;
 typedef uint32_t gid_t;
 typedef uint8_t cpu_id_t;
-typedef uint64_t task_time_t;  // 改名避免与系统 time_t 冲突
+typedef uint32_t task_time_t;  // ⚠️ 必须与内核 time_t 定义一致（32位！）
 
 #define OFFSETOF(type, member) ((unsigned long)&((type *)0)->member)
